Move ExtendedKey to key.py
